+++ skeleton-app/src/lib/components/organisms/PlayerInfo.svelte
